pyproject.toml
[project]
name = "steam-price"
version = "1.0.0"
description = "Fetches Steam game prices and converts them between currencies"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name"}
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
# Enable pycodestyle (E), Pyflakes (F), isort (I), 
# mccabe (C90), comprehensions (C4), builtins (B), 
# bandit (S), and pylint (PL)
select = ["E", "F", "I", "C90", "C4", "B", "S", "PL"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint.mccabe]
# Flag functions with a complexity higher than 12
max-complexity = 12

[tool.ruff.lint.isort]
known-third-party = ["requests", "dotenv"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.pylint]
max-args = 8
max-statements = 50

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true